% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{The Birthday Function}
\usage{
birthday(n)
}
\arguments{
\item{n}{The number of students (can be single value or vector)}
}
\value{
A scatter plot with marginal histogram.
}
\description{
The Birthday Function
}
\examples{
birthday(5)
}
