% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{DISTANCE}{The distance}
  \item{DAMAGE}{The damage done}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
fire
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
