% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My Boot (2nd Iteration)}
\usage{
myboot2(x, iter = 10000, fun = "mean", alpha = 0.05, pop_mean, ...)
}
\arguments{
\item{x}{A vector containing simulated samples}

\item{iter}{The number of iterations for the function to run}

\item{fun}{What statistic is being calculated (e.g. "mean")}

\item{alpha}{The alpha value in a confidence interval (e.g. (1-alpha)*100\%)}

\item{pop_mean}{The mean of the population}

\item{...}{Other values that need to be added}
}
\value{
A histogram and statistics to the console.
}
\description{
Simulation of bootstrap
}
\examples{
myboot2(x = c(5, 7, 9), iter=10000, fun="mean", alpha = 0.05, pop_mean = 25)
}
