% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My N Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The Mean}

\item{sigma}{The standard deviation}

\item{a}{The upper limit for the probability}
}
\value{
a curve graph with the given probability shaded in
}
\description{
My N Curve
}
\examples{
myncurve(5, 10, 7)
}
