% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{The Airline Seat Booking Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{The number of seats available on the plane}

\item{gamma}{The probability of overbooking}

\item{p}{The probability the passenger will show}
}
\value{
Console output of the nd, nc, N, p, and gamma variables
}
\description{
The Airline Seat Booking Function
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)
}
