% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R, R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{A scatter plot with marginal histograms}
\usage{
scatterhist(x, y, xlab = "", ylab = "")

scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{The independent variables to be compared.}

\item{y}{The dependent variables to be compared}

\item{xlab}{Text label for independent variables}

\item{ylab}{Text label for dependent variables}
}
\value{
A scatter plot of x vs y with marginal histograms

A scatter plot with marginal histogram.
}
\description{
A scatter plot with marginal histograms

Scatter Histogram
}
\examples{
scatterhist(x = 1:10, y = 11:20)
scatterhist(c(1, 2, 3, 4), c(2, 4, 6, 8), xlab="Number of calories", ylab="Calories consumed")
}
