% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timesample.R
\name{timesample}
\alias{timesample}
\title{Timed Sample Charts}
\usage{
timesample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{A vector with a sample}

\item{iter}{number of bar charts to make}

\item{time}{Time (in seconds) between each bar to be displayed.}
}
\value{
Generates a new bar chart to console every specified second.
}
\description{
Timed Sample Charts
}
\examples{
timesample(100, 10, 1)
}
