% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{My Central Limit Poisson}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Sample size}

\item{iter}{The number of iterations}

\item{lambda}{lambda parameter for poisson}

\item{...}{Other arguments to be sent to the histogram}
}
\value{
Histogram and Barplot
}
\description{
My Central Limit Poisson
}
\examples{
mycltp(100, 60, 6)
}
